% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyQuestion.R
\name{as.SurveyQuestion}
\alias{as.SurveyQuestion}
\alias{as.SurveyQuestion.data.frame}
\title{Coerce to SurveyQuestion}
\usage{
as.SurveyQuestion(x, ...)

\method{as.SurveyQuestion}{data.frame}(x, questionProperties)
}
\arguments{
\item{x}{Data frame containing all of the columns described in
\link{SurveyQuestion} to be coerced into a SurveyQuestion object}

\item{questionProperties}{Data frame defining the properties of each question
in the survey. See \link{questionProperties} for structure and details}
}
\description{
Functions to check if an object is a valid SurveyQuestion, or coerce it if
possible
}
\details{
Test link \code{\link{SurveyQuestion}}
}

