% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSurveyMonkeyXLS.R
\name{loadSurveyMonkeyXLS}
\alias{loadSurveyMonkeyXLS}
\title{Read an XLS format export from SurveyMonkey.}
\usage{
loadSurveyMonkeyXLS(fname, idcols = 1:9)
}
\arguments{
\item{fname}{Character vector of length 1 pointing to file location}

\item{idcols}{Numeric vector identifying columns in the data to be treated
as respondent identifiers as opposed to question responses.}
}
\value{
data frame of class SurveyQuestion. Each row represents a response
}
\description{
This function parses an XLS-format exported file from SurveyMonkey into a
data frame. The function attempts to deduce the format of each survey
question (e.g. multiple versus single response) from properties of the
headers and the responses in the file.
}
\details{
To obtain the proper file from SurveyMonkey, request a data export of
"all responses" and choose the XLS format (\strong{not} XLS+).
}

