% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonkeyWrangler.R, R/monkeywrangler-package.r
\docType{package}
\name{monkeywrangler}
\alias{monkeywrangler}
\alias{monkeywrangler-package}
\title{monkeywrangler: Read, Visualize, and Summarize Survey Monkey Results}
\description{
monkeywrangler: Read, Visualize, and Summarize Survey Monkey Results

monkeywrangler.
}

