% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyQuestion.R
\name{SurveyQuestion}
\alias{SurveyQuestion}
\title{A class for representing parsed survey responses}
\usage{
SurveyQuestion(respondentId, questionId, question, subgroup, response, type,
  questionProperties)
}
\arguments{
\item{respondentId}{Factor of identifiers to link individual response rows
from the same individual respondent.}

\item{questionId}{Factor of identifiers to link individual response rows from
the same survey question.}

\item{question}{Factor of the text of the questions from the survey
corresponding to the \code{response}.}

\item{subgroup}{Factor of the text of items when the survey question was of a
block or matrix format with responses given to a number of different items}

\item{response}{Character vector of the individual responses given in the
survey}

\item{type}{Factor describing the format of the corresponding question. The
format of a question determines how it is treated by plot and summary
methods}

\item{questionProperties}{Data frame defining the properties of each question
in the survey. See \link{questionProperties} for structure and details}
}
\description{
This long format data frame is produced by \link{loadSurveyMonkeyXLS}.
Methods are defined for \link[=summary.SurveyQuestion]{summary} and
\link[=plot.SurveyQuestion]{plot}. It is unlikely that you will need to call
constructor method directly. However, this help file describes the structure
of the SurveyQuestion object.
}
\details{
The questionProperties object is stored as an attribute named
"questionProperties" in the SurveyQuestion object. It contains one row for
each question in the survey. It contains \code{questionId},
\code{question}, and \code{type} columns similar to those
described above as well as two list columns, \code{subgroups} and
\code{responses}, which store the factor levels that apply to each individual
question. The
\code{questionId}, \code{question}, and \code{type} columns are duplicated in
the main SurveyQuestion data frame for human readability, but all methods
use the values stored in the questionProperties attribute. Therefore, if you
need to alter the properties of a question, you will need make those changes
in the questionProperties attributes. Convenience functions exist for common
changes.
}
\seealso{
\code{\link{questionProperties}}, \code{\link{loadSurveyMonkeyXLS}},
\code{\link{as.SurveyQuestion}}
}

