% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyQuestion.R
\name{getQProps}
\alias{getQProps}
\alias{getResponseLevels}
\alias{setQProps}
\alias{setResponseLevels}
\title{Retreive and Modify Survey Question Properties}
\usage{
getQProps(x)

setQProps(x, questionProperties)

setResponseLevels(x, questionId, levels, var = c("responses", "subgroups"),
  ordinal = F, onlyMatches = F)

getResponseLevels(x, questionId, var = c("responses", "subgroups"))
}
\arguments{
\item{x}{\code{\link{SurveyQuestion}} object.}

\item{questionProperties}{\code{\link{questionProperties}} object.}

\item{questionId}{Integer vector identifying one or more questions for
retrieval or updating}

\item{levels}{Character vector or named list desrcibing how to restructure
the factor levels. See details.}

\item{var}{Character identifying whether to update/retrieve the levels for
the \code{response} or \code{subgroup} columns in \code{x}}

\item{ordinal}{Logical indicating whether the levels should be treated as an
ordered scale. Alters the way that population estimates are determined in
plot and summary methods.}

\item{onlyMatches}{Logical. Should levels only be updated for questions where
each existing level matches with an item in \code{levels}? See details.}
}
\value{
For \code{getQProps}, a data frame wiht columns questionId (integer),
  question (factor), type (factor), subgroups (list), and responses (list).
  For \code{setQProps} and \code{setResponseLevels}, an updated
  \code{\link{SurveyQuestion}} object. For \code{getResponseLevels}, a
  character vector or list in the same format as \code{levels} above,
  optinally containing an "ordinal" attribute, or a list of such when
  \code{questionId} has length greater than 1.
}
\description{
Functions to access or modify the information stored in the
\code{\link{questionProperties}} attribute of \code{\link{SurveyQuestion}}
objects. The most commonly used of these is \code{setResponseLevels}, which
can rearrange or relabel the response options from multiple choice survey
questions and identify ordinal survey scales (e.g., Likert), in order to
generate appropriate population estimates in plots and summaries.
}
\section{Functions}{
\itemize{
\item \code{setQProps}: Replaces the \code{questionProperties} attribute of
\code{x}. No validity checks are performed and no propagation of changes
into the \code{SurveyQuestion} data occurs, so the use of this function to
modify question properties discouraged. One normal use is restoring the
\code{questionProperties} attribute after utilizing functions such as
\code{\link[dplyr]{mutate}} which cause attributes to be lost, having first
preserved the \code{questionProperties} by storing the result of
\code{getQProps} in a variable.

\item \code{setResponseLevels}: The SurveyMonkey XLS export contains limited
information about the original response options. For single repsonse
questions, options that were never selected will be omitted and their order
will not be retained. For both single and multiple response questions, any
"weighting" or ordinal ranking of options is never retained. The plot and
summary methods in this package can make use of that information, but it
needs to be restored with \code{setResponseLevels}. This function is
designed to be versatile to make that process as simple as possible.
Changes made via \code{setResponseLevels} do not alter the underlying data
in the \code{SurveyQuestion} object; instead they alter how that data is
provided when other methods use \code{\link{extractQuestion}} to
access it.

\item \code{getResponseLevels}: temp
}}
\section{Character v. List \code{levels}}{

 When \code{levels} is provided as a
  character vecetor, it is passed to the \code{factor} or \code{ordered}
  constructor via its \code{levels} argument.
}

