% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSurveyMonkeyXLS.R
\name{loadSurveyMonkeyXLS}
\alias{loadSurveyMonkeyXLS}
\title{Read an XLS format export from SurveyMonkey.}
\usage{
loadSurveyMonkeyXLS(fname, idcols = 1:9)
}
\arguments{
\item{fname}{Character vector of length 1 pointing to file location.}

\item{idcols}{Numeric vector identifying columns in the data to be treated as
respondent identifiers as opposed to question responses.}
}
\value{
an object of class \code{\link{SurveyQuestion}} that inherits from
  data.frame. The data is in long form wherein each row respresents a single
  selection from the with the corresponding question named in the
  \code{question} column and the selected or entered value in the
  \code{response} column. The columns specified in the \code{idcols} argument
  will appear in data as well as a unique \code{questionId}, question format
  \code{type}, and \code{subgroup} within a survey question where
  appropriate. Methods are available for
  \link[=summary.SurveyQuestion]{summary} and
  \link[=plot.SurveyQuestion]{plot} for the entire survey or for single
  questions retrieved via \link{extractQuestion}.
}
\description{
Parse an XLS-format exported file from SurveyMonkey into a data frame. The
function attempts to deduce the format of each survey question (e.g. multiple
versus single response) from patterns in the headers and the responses in the
file.
}
\details{
To obtain the proper file from SurveyMonkey, request a data export of "all
responses" and choose the XLS format (\strong{not} XLS+).
}
\seealso{
\code{\link{SurveyQuestion}}, \code{\link{extractQuestion}},
  \code{\link{plot.SurveyQuestion}}, \code{\link{summary.SurveyQuestion}}
}

